pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weaviate-vibe-eval"
version = "0.1.0"
description = "Evaluate LLM capabilities with Weaviate client code"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.22.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["weaviate_vibe_eval"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
